finish main index impl
