+++ pixels-index/pixels-index-rocksdb/src/main/java/io/pixelsdb/pixels/index/rocksdb/RocksDBIndex.java
@@ -221,7 +221,7 @@ public class RocksDBIndex implements SinglePointIndex
             {
                 // Extract key and rowId from Entry object
                 IndexProto.IndexKey key = entry.getIndexKey();
-                long rowId = entry.getTableRowId();
+                long rowId = entry.getRowId();
                 // Convert IndexKey to byte array
                 byte[] keyBytes = toByteArray(key);
                 // Convert rowId to byte array
@@ -249,7 +249,7 @@ public class RocksDBIndex implements SinglePointIndex
             {
                 // Extract key and rowId from Entry object
                 IndexProto.IndexKey key = entry.getIndexKey();
-                long rowId = entry.getTableRowId();
+                long rowId = entry.getRowId();
                 boolean unique = entry.getUnique();
                 // Convert IndexKey to byte array
                 byte[] keyBytes = toByteArray(key);
